example/requirements.txt
flask
werkzeug
jinja2
peewee
wtforms
wtf-peewee
Flask-PyMongo
Flask-Assets
